New shortned placeholderapi IslandRate prefix & other fixes - 13.3.6c
This update mainly adds a new optional feature to shorten IslandRate's placeholderapi prefix from 'islandrate' to 'ir'. You can change this to the latter by changing the default 'placeholderapi_shortened' to 'true'.

This update also fixes the '/rate invcheck' command so that it's actually implemented, and removes unused commands and obsolete code.